requirements.txt
future==0.16.0
pefile==2016.3.28
PyInstaller==3.2
pypiwin32==219
python-dateutil==2.6.0
pytz==2016.7
six==1.10.0
tzlocal==1.3
